+++ CHANGELOG.md
@@ -10,7 +10,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Add "encoding" property for bytes
 * Remove use of standard libraries (std::basic_string, std::vector, std::map)?
 * Use explicit allocator for QuantityArray?
-* Add allowSingular for object
 * Add deprecating option for properties
 * Add normalization
 	- units
@@ -32,6 +31,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Continuous integration
 * Validation of circular, conflicting, & singular aliases
 * Normalization of default, aliases, & allowSingular w/ error handling
+* Add allowSingular for object
 
 ## [Unreleased]